+++ pyvrp/stop/StoppingCriterion.py
@@ -6,7 +6,7 @@ class StoppingCriterion(Protocol):  # pragma: no cover
     Protocol that stopping criteria must implement.
     """
 
-    def __call__(self, best_cost: float) -> bool:
+    def __call__(self, best_cost: int) -> bool:
         """
         When called, this stopping criterion should return True if the
         algorithm should stop, and False otherwise.
